Process files from attachment.tsv

+++ bin/documents.sh
@@ -5,13 +5,18 @@
 #
 
 set -e
-find cache/attachment -type f |
-while read path
+cat data/attachment.tsv |
+while read attachment filename rest
 do
-  id=$(echo "$path"|sed -e 's/^cache\/attachment\///' -e 's/\/.*$//')
-  dir="documents/attachment/$id"
+  if [ "$attachment" == "attachment" ] ; then
+      continue
+  fi
+
+  path=cache/attachment/$attachment/$filename
 
   suffix=$(echo ${path##*.} | tr '[:upper:]' '[:lower:]')
+
+  dir="documents/attachment/$attachment"
   doc="$dir/document.$suffix"
   pdf="$dir/document.pdf"
   json="$dir/document.json"
@@ -22,17 +27,22 @@ do
   if [ ! -f "$pdf" ] ; then
     case "$suffix" in
     zip)
-      echo "$id: skipping ZIP"
+      echo "$attachment: skipping ZIP"
       continue
       ;;
 
     pdf)
-      echo "$id: linking PDF"
-      ln -f "$path" "$pdf"
+      if [ -s "$pdf" ] ; then
+        echo "$attachment: linking PDF"
+        ln -f "$path" "$pdf"
+      else
+        echo "$attachment: skipping missing/empty PDF"
+        continue
+      fi
       ;;
 
     *)
-      echo "$id: converting to PDF"
+      echo "$attachment: converting to PDF"
       (
         ln -f "$path" "$doc"
         cd $dir
@@ -44,34 +54,30 @@ do
   fi
 
   if ls $dir/page*.png > /dev/null 2>&1 ; then : ; else
-    echo "$id: extracting images"
+    echo "$attachment: extracting images"
     pdftoppm -l 99 -png "$pdf" "$dir/page"
   fi
 
   if [ ! -f "$json" ] ; then
-    echo "$id: extracting metadata"
+    echo "$attachment: extracting metadata"
     if java -jar cache/tika.jar -j "$path" > "$json" 2> "$dir/tika-json.err" ; then
       rm $dir/tika-json.err
     fi
   fi
 
   if [ ! -f "$txt" ] ; then
-    echo "$id: extracting text"
+    echo "$attachment: extracting text"
 
     # gs -q -sDEVICE=txtwrite -o "$dir/page-%d.txt" "$pdf"
     # pdftotext "$pdf"
 
-    if java -jar cache/tika.jar -t "$path" > "$txt" 2> "$dir/tika-txt.err" ; then
-      rm $dir/tika-txt.err
-      tmp=/tmp/tika.$$
-
-      # reduce blanks ..
-      sed \
-        -e 's/[[:space:]]*$//'\
-        -e '/^$/N;/^\n$/D' \
-        < "$txt" > "$tmp"
-
-      mv "$tmp" "$txt"
+    tmp=$txt.$$
+    if java -jar cache/tika.jar -t "$path" > "$tmp" 2> "$dir/tika-txt.err"
+    then
+      python3 bin/blanks.py < "$tmp" > "$txt"
+      rm $dir/tika-txt.err $tmp
+    else
+      echo "$attachment: tika failed for $path"
     fi
   fi
 done
